# docs/openapi.yml
openapi: 3.0.3
info:
  title: Leave Management System API
  description: |
    A comprehensive leave management system backend built with Node.js, Express.js, and Supabase.
    This API provides endpoints for managing employee leave requests, approvals, and balance tracking.
    
    ## Features
    - Employee management with leave balance tracking
    - Leave application with comprehensive validation
    - Manager approval workflow with business rules
    - Real-time leave balance calculations
    - Audit trail and reporting capabilities
    
    ## Authentication
    This API uses Supabase authentication. Include your authentication token in the Authorization header.
    
    ## Rate Limiting
    API requests are limited to 100 requests per minute per IP address.
    
    ## Error Handling
    All errors follow RFC 7807 problem details format with appropriate HTTP status codes.
  version: 1.0.0
  contact:
    name: Leave Management System Support
    email: support@company.com
    url: https://github.com/Abhinav12004/Leave-Management-System
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://company.com/terms
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://staging-api.company.com
    description: Staging server
  - url: https://api.company.com
    description: Production server
paths:
...
  /employees:
    get:
      summary: Get all employees
      description: Retrieve a list of all employees with their basic information
      tags:
        - Employees
      responses:
        '200':
          description: Employees retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Employees retrieved successfully"
                  employees:
                    type: array
                    items:
                      $ref: '#/components/schemas/Employee'
                  count:
                    type: integer
                    example: 5
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Create a new employee
      description: Create a new employee with the provided information
      tags:
        - Employees
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeRequest'
      responses:
        '201':
          description: Employee created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Employee created successfully"
                  employee:
                    $ref: '#/components/schemas/Employee'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Employee:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john.doe@company.com"
        department:
          type: string
          example: "Engineering"
        joining_date:
          type: string
          format: date
          example: "2024-01-15"
        leave_balance:
          type: integer
          example: 20
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateEmployeeRequest:
      type: object
      required:
        - name
        - email
        - department
        - joining_date
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: "John Doe"
          description: Employee full name
        email:
          type: string
          format: email
          maxLength: 255
          example: "john.doe@company.com"
          description: Unique email address
        department:
          type: string
          enum: [Engineering, Human Resources, Finance, Marketing, Sales, Operations, Legal, Management, IT, HR]
          example: "Engineering"
          description: Employee department
        joining_date:
          type: string
          format: date
          example: "2024-01-15"
          description: Employee joining date (YYYY-MM-DD)
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Something went wrong"
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          example: "abc123def456"
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            validation_errors:
              type: array
              items:
                type: string
              example: ["Missing required field: start_date", "Invalid email format"]
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT token
security:
  - bearerAuth: []
...